# Fix #44 Add support for circular getter references