--- a/benchmark.js
+++ b/benchmark.js
@@ -8,6 +8,9 @@ const array = new Array(10).fill(0).map((_, i) => i)
 const obj = { foo: array }
 const circ = JSON.parse(JSON.stringify(obj))
 circ.o = { obj: circ, array }
+const circGetters = JSON.parse(JSON.stringify(obj))
+Object.assign(circGetters, { get o () { return { obj: circGetters, array } } })
+
 const deep = require('./package.json')
 deep.deep = JSON.parse(JSON.stringify(deep))
 deep.deep.deep = JSON.parse(JSON.stringify(deep))
@@ -20,43 +23,75 @@ deepCirc.deep.deep.circ = deepCirc
 deepCirc.deep.circ = deepCirc
 deepCirc.array = array
 
-suite.add('util.inspect:          simple object', function () {
-  inspect(obj)
+const deepCircGetters = JSON.parse(JSON.stringify(deep))
+for (let i = 0; i < 10; i++) {
+  deepCircGetters[i.toString()] = {
+    deep: {
+      deep: {
+        get circ () { return deep.deep },
+        deep: { get circ () { return deep.deep.deep } }
+      },
+      get circ () { return deep }
+    },
+    get array () { return array }
+  }
+}
+
+suite.add('util.inspect:          simple object            ', function () {
+  inspect(obj, { showHidden: false, depth: null })
 })
-suite.add('util.inspect:          circular     ', function () {
-  inspect(circ)
+suite.add('util.inspect:          circular                 ', function () {
+  inspect(circ, { showHidden: false, depth: null })
 })
-suite.add('util.inspect:          deep         ', function () {
-  inspect(deep)
+suite.add('util.inspect:          circular getters         ', function () {
+  inspect(circGetters, { showHidden: false, depth: null })
 })
-suite.add('util.inspect:          deep circular', function () {
-  inspect(deepCirc)
+suite.add('util.inspect:          deep                     ', function () {
+  inspect(deep, { showHidden: false, depth: null })
+})
+suite.add('util.inspect:          deep circular            ', function () {
+  inspect(deepCirc, { showHidden: false, depth: null })
+})
+suite.add('util.inspect:          10+ deep circular getters', function () {
+  inspect(deepCircGetters, { showHidden: false, depth: null })
 })
 
-suite.add('\njson-stringify-safe:   simple object', function () {
+suite.add('\njson-stringify-safe:   simple object            ', function () {
   jsonStringifySafe(obj)
 })
-suite.add('json-stringify-safe:   circular     ', function () {
+suite.add('json-stringify-safe:   circular                 ', function () {
   jsonStringifySafe(circ)
 })
-suite.add('json-stringify-safe:   deep         ', function () {
+suite.add('json-stringify-safe:   circular getters         ', function () {
+  jsonStringifySafe(circGetters)
+})
+suite.add('json-stringify-safe:   deep                     ', function () {
   jsonStringifySafe(deep)
 })
-suite.add('json-stringify-safe:   deep circular', function () {
+suite.add('json-stringify-safe:   deep circular            ', function () {
   jsonStringifySafe(deepCirc)
 })
+suite.add('json-stringify-safe:   10+ deep circular getters', function () {
+  jsonStringifySafe(deepCircGetters)
+})
 
-suite.add('\nfast-safe-stringify:   simple object', function () {
+suite.add('\nfast-safe-stringify:   simple object            ', function () {
   fastSafeStringify(obj)
 })
-suite.add('fast-safe-stringify:   circular     ', function () {
+suite.add('fast-safe-stringify:   circular                 ', function () {
   fastSafeStringify(circ)
 })
-suite.add('fast-safe-stringify:   deep         ', function () {
+suite.add('fast-safe-stringify:   circular getters         ', function () {
+  fastSafeStringify(circGetters)
+})
+suite.add('fast-safe-stringify:   deep                     ', function () {
   fastSafeStringify(deep)
 })
-suite.add('fast-safe-stringify:   deep circular', function () {
+suite.add('fast-safe-stringify:   deep circular            ', function () {
   fastSafeStringify(deepCirc)
+})
+suite.add('fast-safe-stringify:   10+ deep circular getters', function () {
+  fastSafeStringify(deepCircGetters)
 })
 
 // add listeners