--- a/benchmark.js
+++ b/benchmark.js
@@ -59,69 +59,69 @@ Object.defineProperty(deepCircNonCongifurableGetters, 'array', {
   configurable: false
 })
 
-suite.add('util.inspect:          simple object              ', function () {
+suite.add('util.inspect:          simple object                 ', function () {
   inspect(obj, { showHidden: false, depth: null })
 })
-suite.add('util.inspect:          circular                   ', function () {
+suite.add('util.inspect:          circular                      ', function () {
   inspect(circ, { showHidden: false, depth: null })
 })
-suite.add('util.inspect:          circular getters           ', function () {
+suite.add('util.inspect:          circular getters              ', function () {
   inspect(circGetters, { showHidden: false, depth: null })
 })
-suite.add('util.inspect:          deep                       ', function () {
+suite.add('util.inspect:          deep                          ', function () {
   inspect(deep, { showHidden: false, depth: null })
 })
-suite.add('util.inspect:          deep circular              ', function () {
+suite.add('util.inspect:          deep circular                 ', function () {
   inspect(deepCirc, { showHidden: false, depth: null })
 })
-suite.add('util.inspect:          large deep circular getters', function () {
+suite.add('util.inspect:          large deep circular getters   ', function () {
   inspect(deepCircGetters, { showHidden: false, depth: null })
 })
-suite.add('util.inspect:          deepCircNonCongifurableGetters', function () {
+suite.add('util.inspect:          deep non-conf circular getters', function () {
   inspect(deepCircNonCongifurableGetters, { showHidden: false, depth: null })
 })
 
-suite.add('\njson-stringify-safe:   simple object              ', function () {
+suite.add('\njson-stringify-safe:   simple object                 ', function () {
   jsonStringifySafe(obj)
 })
-suite.add('json-stringify-safe:   circular                   ', function () {
+suite.add('json-stringify-safe:   circular                      ', function () {
   jsonStringifySafe(circ)
 })
-suite.add('json-stringify-safe:   circular getters           ', function () {
+suite.add('json-stringify-safe:   circular getters              ', function () {
   jsonStringifySafe(circGetters)
 })
-suite.add('json-stringify-safe:   deep                       ', function () {
+suite.add('json-stringify-safe:   deep                          ', function () {
   jsonStringifySafe(deep)
 })
-suite.add('json-stringify-safe:   deep circular              ', function () {
+suite.add('json-stringify-safe:   deep circular                 ', function () {
   jsonStringifySafe(deepCirc)
 })
-suite.add('json-stringify-safe:   large deep circular getters', function () {
+suite.add('json-stringify-safe:   large deep circular getters   ', function () {
   jsonStringifySafe(deepCircGetters)
 })
-suite.add('json-stringify-safe:   deepCircNonCongifurableGetters', function () {
+suite.add('json-stringify-safe:   deep non-conf circular getters', function () {
   jsonStringifySafe(deepCircNonCongifurableGetters)
 })
 
-suite.add('\nfast-safe-stringify:   simple object              ', function () {
+suite.add('\nfast-safe-stringify:   simple object                 ', function () {
   fastSafeStringify(obj)
 })
-suite.add('fast-safe-stringify:   circular                   ', function () {
+suite.add('fast-safe-stringify:   circular                      ', function () {
   fastSafeStringify(circ)
 })
-suite.add('fast-safe-stringify:   circular getters           ', function () {
+suite.add('fast-safe-stringify:   circular getters              ', function () {
   fastSafeStringify(circGetters)
 })
-suite.add('fast-safe-stringify:   deep                       ', function () {
+suite.add('fast-safe-stringify:   deep                          ', function () {
   fastSafeStringify(deep)
 })
-suite.add('fast-safe-stringify:   deep circular              ', function () {
+suite.add('fast-safe-stringify:   deep circular                 ', function () {
   fastSafeStringify(deepCirc)
 })
-suite.add('fast-safe-stringify:   large deep circular getters', function () {
+suite.add('fast-safe-stringify:   large deep circular getters   ', function () {
   fastSafeStringify(deepCircGetters)
 })
-suite.add('fast-safe-stringify:   deepCircNonCongifurableGetters', function () {
+suite.add('fast-safe-stringify:   deep non-conf circular getters', function () {
   fastSafeStringify(deepCircNonCongifurableGetters)
 })