--- a/index.js
+++ b/index.js
@@ -20,6 +20,9 @@ function decirc (val, k, stack, parent) {
   if (typeof val === 'object' && val !== null) {
     for (i = 0; i < stack.length; i++) {
       if (stack[i] === val) {
+        if (typeof Object.getOwnPropertyDescriptor(parent, k).get === 'function') {
+          delete parent[k]
+        }
         parent[k] = '[Circular]'
         arr.push([parent, k, val])
         return