--- a/index.js
+++ b/index.js
@@ -4,11 +4,17 @@ stringify.stable = deterministicStringify
 stringify.stableStringify = deterministicStringify
 
 var arr = []
+var nonConfigurables = []
 
 // Regular stringify
 function stringify (obj, replacer, spacer) {
   decirc(obj, '', [], undefined)
-  var res = JSON.stringify(obj, replacer, spacer)
+  var res
+  if (nonConfigurables.length === 0) {
+    res = JSON.stringify(obj, replacer, spacer)
+  } else {
+    res = JSON.stringify(obj, handleNonConfigurableGetters(replacer), spacer)
+  }
   while (arr.length !== 0) {
     var part = arr.pop()
     if (part.length === 4) {
@@ -24,10 +30,15 @@ function decirc (val, k, stack, parent) {
   if (typeof val === 'object' && val !== null) {
     for (i = 0; i < stack.length; i++) {
       if (stack[i] === val) {
+        // Note: This doesn't handle circular sub-objects of non-configurable getters...
         const propertyDescriptor = Object.getOwnPropertyDescriptor(parent, k)
         if (propertyDescriptor.get !== undefined) {
-          Object.defineProperty(parent, k, { value: '[Circular]' })
-          arr.push([parent, k, val, propertyDescriptor])
+          if (propertyDescriptor.configurable) {
+            Object.defineProperty(parent, k, { value: '[Circular]' })
+            arr.push([parent, k, val, propertyDescriptor])
+          } else {
+            nonConfigurables.push([parent, k])
+          }
         } else {
           parent[k] = '[Circular]'
           arr.push([parent, k, val])
@@ -65,7 +76,12 @@ function compareFunction (a, b) {
 
 function deterministicStringify (obj, replacer, spacer) {
   var tmp = deterministicDecirc(obj, '', [], undefined) || obj
-  var res = JSON.stringify(tmp, replacer, spacer)
+  var res
+  if (nonConfigurables.length === 0) {
+    res = JSON.stringify(tmp, replacer, spacer)
+  } else {
+    res = JSON.stringify(tmp, handleNonConfigurableGetters(replacer), spacer)
+  }
   while (arr.length !== 0) {
     var part = arr.pop()
     if (part.length === 4) {
@@ -82,10 +98,15 @@ function deterministicDecirc (val, k, stack, parent) {
   if (typeof val === 'object' && val !== null) {
     for (i = 0; i < stack.length; i++) {
       if (stack[i] === val) {
+        // Note: This doesn't handle circular sub-objects of non-configurable getters...
         const propertyDescriptor = Object.getOwnPropertyDescriptor(parent, k)
         if (propertyDescriptor.get !== undefined) {
-          Object.defineProperty(parent, k, { value: '[Circular]' })
-          arr.push([parent, k, val, propertyDescriptor])
+          if (propertyDescriptor.configurable) {
+            Object.defineProperty(parent, k, { value: '[Circular]' })
+            arr.push([parent, k, val, propertyDescriptor])
+          } else {
+            nonConfigurables.push([parent, k])
+          }
         } else {
           parent[k] = '[Circular]'
           arr.push([parent, k, val])
@@ -119,5 +140,25 @@ function deterministicDecirc (val, k, stack, parent) {
       }
     }
     stack.pop()
+  }
+}
+
+// wraps replacer function to handle non-configurable getters
+// and sort of mark them as [Circular]
+function handleNonConfigurableGetters (replacer) {
+  replacer = replacer !== undefined ? replacer : (k, v) => v
+  return function (key, val) {
+    if (nonConfigurables.length > 0) {
+      for (let i = 0; i < nonConfigurables.length; i++) {
+        const part = nonConfigurables[i]
+        // This references the parent object being stringified
+        if (part[1] === key && part[0] === this) {
+          val = '[Circular]'
+          nonConfigurables.splice(i, 1)
+          break
+        }
+      }
+    }
+    return replacer.call(this, key, val)
   }
 }