--- a/index.js
+++ b/index.js
@@ -4,16 +4,16 @@ stringify.stable = deterministicStringify
 stringify.stableStringify = deterministicStringify
 
 var arr = []
-var nonConfigurables = []
+var replacerStack = []
 
 // Regular stringify
 function stringify (obj, replacer, spacer) {
   decirc(obj, '', [], undefined)
   var res
-  if (nonConfigurables.length === 0) {
+  if (replacerStack.length === 0) {
     res = JSON.stringify(obj, replacer, spacer)
   } else {
-    res = JSON.stringify(obj, handleNonConfigurableGetters(replacer), spacer)
+    res = JSON.stringify(obj, replaceGetterValues(replacer), spacer)
   }
   while (arr.length !== 0) {
     var part = arr.pop()
@@ -30,14 +30,13 @@ function decirc (val, k, stack, parent) {
   if (typeof val === 'object' && val !== null) {
     for (i = 0; i < stack.length; i++) {
       if (stack[i] === val) {
-        // Note: This doesn't handle circular sub-objects of non-configurable getters...
         var propertyDescriptor = Object.getOwnPropertyDescriptor(parent, k)
         if (propertyDescriptor.get !== undefined) {
           if (propertyDescriptor.configurable) {
             Object.defineProperty(parent, k, { value: '[Circular]' })
             arr.push([parent, k, val, propertyDescriptor])
           } else {
-            nonConfigurables.push([parent, k])
+            replacerStack.push([val, k])
           }
         } else {
           parent[k] = '[Circular]'
@@ -77,10 +76,10 @@ function compareFunction (a, b) {
 function deterministicStringify (obj, replacer, spacer) {
   var tmp = deterministicDecirc(obj, '', [], undefined) || obj
   var res
-  if (nonConfigurables.length === 0) {
+  if (replacerStack.length === 0) {
     res = JSON.stringify(tmp, replacer, spacer)
   } else {
-    res = JSON.stringify(tmp, handleNonConfigurableGetters(replacer), spacer)
+    res = JSON.stringify(tmp, replaceGetterValues(replacer), spacer)
   }
   while (arr.length !== 0) {
     var part = arr.pop()
@@ -98,14 +97,13 @@ function deterministicDecirc (val, k, stack, parent) {
   if (typeof val === 'object' && val !== null) {
     for (i = 0; i < stack.length; i++) {
       if (stack[i] === val) {
-        // Note: This doesn't handle circular sub-objects of non-configurable getters...
         var propertyDescriptor = Object.getOwnPropertyDescriptor(parent, k)
         if (propertyDescriptor.get !== undefined) {
           if (propertyDescriptor.configurable) {
             Object.defineProperty(parent, k, { value: '[Circular]' })
             arr.push([parent, k, val, propertyDescriptor])
           } else {
-            nonConfigurables.push([parent, k])
+            replacerStack.push([val, k])
           }
         } else {
           parent[k] = '[Circular]'
@@ -143,18 +141,17 @@ function deterministicDecirc (val, k, stack, parent) {
   }
 }
 
-// wraps replacer function to handle non-configurable getters
-// and sort of mark them as [Circular]
-function handleNonConfigurableGetters (replacer) {
-  replacer = replacer !== undefined ? replacer : (k, v) => v
+// wraps replacer function to handle values we couldn't replace
+// and mark them as [Circular]
+function replaceGetterValues (replacer) {
+  replacer = replacer !== undefined ? replacer : function (k, v) { return v }
   return function (key, val) {
-    if (nonConfigurables.length > 0) {
-      for (var i = 0; i < nonConfigurables.length; i++) {
-        var part = nonConfigurables[i]
-        // This references the parent object being stringified
-        if (part[1] === key && part[0] === this) {
+    if (replacerStack.length > 0) {
+      for (var i = 0; i < replacerStack.length; i++) {
+        var part = replacerStack[i]
+        if (part[1] === key && part[0] === val) {
           val = '[Circular]'
-          nonConfigurables.splice(i, 1)
+          replacerStack.splice(i, 1)
           break
         }
       }