@davidmarkclements 2019-08-12
@kristofferostlund thanks for much for this PR, really appreciate you highlighting this

1. I agree the getter should be replaced with `[Circular]` in the output (also we should avoid `delete` as that *will* put the object into slow property mode)
2. I agree decirc should be synced
3. I think we need benchmarks for objects with lots of properties (and lots of get properties) to really see the impact of this change