@kristofferostlund 2019-08-13
Sorry, I should have been more explicit, I did mark getters in `decirc` as `[Circular]`.

I tried different ways of marking the getters as `[Circular]` and found `Object.defineProperty(parent, k, { 
value: '[Circular]' })` to be most efficient. 

I tried the following other approaches and found them all marginally slower:

```javascript
// Redefining the getter
if (Object.getOwnPropertyDescriptor(parent, k).get !== undefined) {
  Object.defineProperty(parent, k, { get: () => '[Circular]' })
} else {
  parent[k] = '[Circular]'
}

// Un-getterify and set the value
if (Object.getOwnPropertyDescriptor(parent, k).get !== undefined) {
  Object.defineProperty(parent, k, { value: undefined, writable: true, enumerable: true })
}
parent[k] = '[Circular]'

// Deleting the property before marking it as [Circular]
if (Object.getOwnPropertyDescriptor(parent, k).get !== undefined) {
  delete parent[k]
}
parent[k] = '[Circular]'

// And then different ways of checking if it is a getter such as
// checking if the get is a function
if (typeof Object.getOwnPropertyDescriptor(parent, k).get === 'function') {
  ...
}

// Or checking if it's writable, though this could trigger a setters functionality
// and potentially break stuff
if (!Object.getOwnPropertyDescriptor(parent, k).writable) {
  ...
}
```

---

I had to modify the benchmarks when adding larger objects with circular getters since `util.inspect()` doesn't go the full depth ([and different Node version it seems to go to different depths by defualt](https://nodejs.org/api/util.html#util_util_inspect_object_options)) and thus rendered unfair benchmarks:

```
util.inspect:          simple object               x 43,694 ops/sec ±1.06% (90 runs sampled)
util.inspect:          circular                    x 22,233 ops/sec ±0.66% (92 runs sampled)
util.inspect:          circular getters            x 21,869 ops/sec ±0.72% (92 runs sampled)
util.inspect:          deep                        x 1,132 ops/sec ±0.69% (93 runs sampled)
util.inspect:          deep circular               x 1,126 ops/sec ±0.54% (92 runs sampled)
util.inspect:          large deep circular getters x 851 ops/sec ±0.55% (94 runs sampled)

json-stringify-safe:   simple object               x 346,326 ops/sec ±1.24% (91 runs sampled)
json-stringify-safe:   circular                    x 153,442 ops/sec ±0.61% (93 runs sampled)
json-stringify-safe:   circular getters            x 153,272 ops/sec ±0.86% (92 runs sampled)
json-stringify-safe:   deep                        x 8,476 ops/sec ±0.69% (91 runs sampled)
json-stringify-safe:   deep circular               x 8,296 ops/sec ±0.85% (92 runs sampled)
json-stringify-safe:   large deep circular getters x 300 ops/sec ±0.96% (88 runs sampled)

fast-safe-stringify:   simple object               x 1,137,839 ops/sec ±0.71% (92 runs sampled)
fast-safe-stringify:   circular                    x 549,733 ops/sec ±0.72% (91 runs sampled)
fast-safe-stringify:   circular getters            x 555,157 ops/sec ±0.70% (92 runs sampled)
fast-safe-stringify:   deep                        x 38,223 ops/sec ±0.65% (92 runs sampled)
fast-safe-stringify:   deep circular               x 37,554 ops/sec ±0.51% (92 runs sampled)
fast-safe-stringify:   large deep circular getters x 1,335 ops/sec ±0.98% (91 runs sampled)

Fastest is
fast-safe-stringify:   simple object
```