@kristofferostlund 2019-08-13
I don't think my changes are causing Travis to fail. It looks like `tap`'s `ESM` transform somehow misses the closing brackets (`}})`) when running Node.js 4.x?