@kristofferostlund 2019-08-13
I've gotten non-configurable circular getters to work now.

But I found an issue where circular references returned not as the main value, but as  a property of an object is not marked as circular since I believe we now mark the getter's return value's value as `[Circular]`, which is problematic, right?

To clarify:

```javascript
// Won't work for now
const o = {
  a: 1,
  get o() {
    return { o: o }
  }
}
```

I'll look into this tomorrow after work again.

---

Also: this time I think I did break Travis on Node.js 4.x haha!