@kristofferostlund 2019-08-25
Sorry for the slow finish here, I've been ill basically all week. I solved circular references in child elements and added some tests for it.

```
util.inspect:          simple object                  x 44,952 ops/sec ±1.15% (88 runs sampled)
util.inspect:          circular                       x 22,454 ops/sec ±0.72% (89 runs sampled)
util.inspect:          circular getters               x 22,529 ops/sec ±0.73% (94 runs sampled)
util.inspect:          deep                           x 1,076 ops/sec ±0.73% (92 runs sampled)
util.inspect:          deep circular                  x 1,066 ops/sec ±0.56% (92 runs sampled)
util.inspect:          large deep circular getters    x 815 ops/sec ±1.99% (92 runs sampled)
util.inspect:          deep non-conf circular getters x 1,071 ops/sec ±0.73% (92 runs sampled)

json-stringify-safe:   simple object                  x 344,785 ops/sec ±0.81% (89 runs sampled)
json-stringify-safe:   circular                       x 152,935 ops/sec ±0.82% (91 runs sampled)
json-stringify-safe:   circular getters               x 153,137 ops/sec ±0.84% (89 runs sampled)
json-stringify-safe:   deep                           x 8,042 ops/sec ±0.67% (94 runs sampled)
json-stringify-safe:   deep circular                  x 7,861 ops/sec ±1.01% (91 runs sampled)
json-stringify-safe:   large deep circular getters    x 285 ops/sec ±0.79% (86 runs sampled)
json-stringify-safe:   deep non-conf circular getters x 7,427 ops/sec ±0.92% (93 runs sampled)

fast-safe-stringify:   simple object                  x 1,152,173 ops/sec ±0.63% (91 runs sampled)
fast-safe-stringify:   circular                       x 559,578 ops/sec ±0.86% (90 runs sampled)
fast-safe-stringify:   circular getters               x 559,467 ops/sec ±0.70% (92 runs sampled)
fast-safe-stringify:   deep                           x 37,123 ops/sec ±1.14% (89 runs sampled)
fast-safe-stringify:   deep circular                  x 36,857 ops/sec ±0.69% (92 runs sampled)
fast-safe-stringify:   large deep circular getters    x 1,313 ops/sec ±0.56% (91 runs sampled)
fast-safe-stringify:   deep non-conf circular getters x 10,669 ops/sec ±0.98% (93 runs sampled)

Fastest is
fast-safe-stringify:   simple object
```